run.py app.py로 구분, unittest 프로토타입 생성
