+++ run.py
@@ -1,24 +1,4 @@
-from flask import Flask, render_template
-import os
-from dotenv import load_dotenv
-
-load_dotenv()
-
-app = Flask(__name__)
-
-@app.route('/')
-def login():
-
-    return render_template('LogInPage.html')
-
-@app.route('/signup')
-def signup():
-    
-    return render_template('Signup.html')
-
-@app.route('/homepage')
-def homepage():
-    return render_template('homepage.html')
+from app import app
 
 if __name__ == "__main__":
     app.run(host='0.0.0.0', debug=True)
